// src/app/model/activity.ts
export interface IActivityItem {
    activityDate: string
    activityId: string
    activityType: number
    campaignId: string
    companyId: any
    createdDate: any
    deleted: number
    description: string
    firstName: string
    id: number
    lastName: string
    updatedDate: any
    userId: any
}